move provider registration to bootstrap
